requirements.txt
# Add PyTorch CPU wheel repo
--extra-index-url https://download.pytorch.org/whl/cpu

# PyTorch and vision (CPU-only)
torch==2.1.0+cpu
torchvision==0.16.0+cpu

# Core libraries
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
pillow==10.3.0

# ML and Transformers
scikit-learn==1.3.2
transformers==4.40.0

# NLP and OCR
spacy==3.7.4
pytesseract==0.3.10
